Work around a bug in the regressor
Apparently something in the regressor is not initialized
when the estimate is run for the first time, causing the
fixation of the parameters to hit a nullptr dereference.

Work around the problem by dry-running the estimation with
all parameters released at first and re-run the calculation
with the fixed parameters afterwards.